backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
docker==7.0.0
GitPython==3.1.40
requests==2.31.0
gunicorn==21.2.0
psutil==5.9.6
python-dotenv==1.0.0
werkzeug==3.0.0
PyGithub==2.3.0
psycopg2-binary==2.9.9
flask-limiter==3.5.0
redis==5.0.1
SQLAlchemy==2.0.23
